Helper function for random number generation from normal distribution, which will be used to initialize weights

+++ src/NeuralNetwork.cpp
@@ -8,14 +8,15 @@ NeuralNetwork::NeuralNetwork(std::vector<int> topology)
 
     this->topology = topology;
 
-    for (int i=0; i<topology.size(); i++){
+    for (int i=0; i<(int)topology.size(); i++){
         this->l = new Layer(topology.at(i));
         this->layers.push_back(this->l);
     }
 
-    for (int i=0; i<layers.size(); i++){
+    // Log
+    for (int i=0; i<(int)layers.size(); i++){
         std::cout << "Layer: " << i << " : " << this->layers.at(i) << std::endl;
-        for (int j=0; j<this->layers.at(i)->getNeuronsofALayer().size(); j++){
+        for (int j=0; j<(int)this->layers.at(i)->getNeuronsofALayer().size(); j++){
             std::cout << "Neuron: " << j << " : " << this->layers.at(i)->getNeuronsofALayer().at(j)->getNeuronVal() << std::endl;
         }
